[WIP] Projects/codes rewrite

+++ Codes/views.py
@@ -1,7 +1,9 @@
 from django.contrib.auth.decorators import login_required
 from django.contrib import messages
 from django.shortcuts import get_object_or_404, render, redirect
-from .models import Project
+from django.db.models import Q
+
+from .models import Project, Code
 from .forms import ProjectForm
 
 
@@ -14,18 +16,15 @@ def addProject(request):
         if form.is_valid():
             title = form.cleaned_data.get('title')
             description = form.cleaned_data.get('description')
-            source_code = form.cleaned_data.get('source_code')
 
             if not title:
                 form.add_error('title', 'Podaj tytuł dla swojego kodu.')
             elif not description:
                 form.add_error('description', 'Dodaj krótki opis o tym czego dotyczy twój kod')
-            elif not source_code:
-                form.add_error('source_code', 'Kolego, nie można dodać kodu bez kodu.')
             else:
                 new_project = form.save(commit=False)
                 new_project.owner = request.user
-                new_project.plagiarism_ratio = 0
+                new_project.parent = None
                 new_project.save()
 
                 messages.success(request, 'Projekt został poprawnie utworzony.')
@@ -45,19 +44,20 @@ def editProject(request, project_id):
     page = 'edit_project'
     project = get_object_or_404(Project, id=project_id)
 
+    if request.user != project.owner:
+        messages.error(request, 'Nie jesteś właścicielem tego projektu.')
+        return redirect('account')
+
     if request.method == 'POST':
         form = ProjectForm(request.POST, instance=project)
         if form.is_valid():
             title = form.cleaned_data.get('title')
             description = form.cleaned_data.get('description')
-            source_code = form.cleaned_data.get('source_code')
 
             if not title:
                 form.add_error('title', 'Podaj tytuł dla swojego kodu.')
             elif not description:
                 form.add_error('description', 'Dodaj krótki opis o tym czego dotyczy twój kod')
-            elif not source_code:
-                form.add_error('source_code', 'Kolego, nie można dodać kodu bez kodu.')
             else:
                 form.save()
                 messages.success(request, 'Zapisano zmiany.')
@@ -70,25 +70,51 @@ def editProject(request, project_id):
     }
     return render(request, 'Projects/add-edit_project.html', context)
 
+
 @login_required(login_url='login')
 def displayMyProject(request, project_id):
     page = 'display_my_project'
     user = request.user
 
     project = get_object_or_404(Project, id=project_id)
+    codes = Code.objects.filter(
+        Q(owner=user) & Q(project=project)
+    ).order_by('-creation_date')
 
     context = {
         'page': page,
         'project': project,
-        'user': user, 
+        'codes': codes,
+        'user': user,
     }
     return render(request, 'Projects/display_project.html', context)
+
+@login_required(login_url='login')
+def display_code(request, code_id):
+    code = get_object_or_404(Code, id=code_id)
+    page: str = 'display_code'
+
+    # check if user is owner of the code
+    if request.user != code.owner:
+        messages.error('Nie jesteś właścicielem tego kodu.')
+        return redirect('account')
+
+    context = {
+        'page': page,
+        'code': code,
+    }
+
+    return render(request, 'Projects/display_code.html', context)
 
 @login_required(login_url='login')
 def deleteProject(request, project_id):
     project = get_object_or_404(Project, id=project_id)
 
+    # check if user is owner of project
+    if request.user != project.owner:
+        messages.error(request, 'Nie jesteś właścicielem tego projektu.')
+        return redirect('account')
+
     project.delete()
     messages.success(request, 'Projekt został usunięty.')
     return redirect('account')
-
